api-nei/pyproject.toml
[tool.poetry]
name = "api-nei"
version = "0.1.0"
description = "NEI Service"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.3"
SQLAlchemy = {extras = ["postgresql-psycopg2binary"], version = "^2.0.25"}
alembic = "^1.13.1"
loguru = "^0.7.2"
Pillow = "^10.2.0"
python-multipart = "^0.0.6"
websockets = "^12.0"
fastapi = "^0.109.1"
orjson = "^3.9.15"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
email-validator = "^2.1.0.post1"
aiosmtplib = "^3.0.1"
jinja2 = "^3.1.3"
httpx = "^0.26.0"
python-magic = "^0.4.27"
authlib = "^1.3.0"
pydantic-settings = "^2.1.0"
pydantic-extra-types = "^2.4.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
flake8 = "^7.0.0"
sqlalchemy-stubs = "^0.4"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
